chore: int64s all the way down

+++ model/Tweet.go
@@ -4,11 +4,11 @@ import "time"
 
 type Tweet struct {
 	// TODO: change these to int64
-	ID       int
+	ID       int64
 	Author   string
 	Content  string
 	Created  time.Time
-	LikeCount int
+	LikeCount int64
 	LikedBySelf bool
 	BookmarkedBySelf bool
 }
